user routes, policies, job and notification

+++ app/Services/PostService.php
@@ -4,8 +4,11 @@ namespace App\Services;
 
 use App\Http\Resources\PostResource;
 use App\Models\Post;
+use App\Models\User;
 use Illuminate\Pagination\LengthAwarePaginator;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Auth;
+use Laravel\Horizon\Exceptions\ForbiddenException;
 
 class PostService implements PostServiceInterface
 {
@@ -30,7 +33,14 @@ class PostService implements PostServiceInterface
 
     public function updatePost(array $data, int $id): Post
     {
+        /** @var User $user */
+        $user = Auth::user();
         $post = Post::findOrFail($id);
+
+        if ($user->cannot('update', $post)) {
+            throw new ForbiddenException(403, 'You are not allowed to update this post');
+        }
+
         $post->update($data);
 
         return $post;
@@ -38,7 +48,15 @@ class PostService implements PostServiceInterface
 
     public function deletePost(int $id): void
     {
+        /** @var User $user */
+        $user = Auth::user();
+
         $post = Post::findOrFail($id);
+
+        if ($user->cannot('update', $post)) {
+            throw new ForbiddenException(403, 'You are not allowed to delete this post');
+        }
+
         $post->delete($post);
     }
 }
